refactor(gauntlet): drop unused multiplier knob, decompose page files, add aggregate tests
Addresses code-review feedback on the gauntlet branch:

- Drop the unused `multiplier` field from GauntletModifier; centralize
  HOT_LETTER_MULTIPLIER in models/gauntlet.ts so engine, server config,
  and display strings share a single source.
- Extract scoreFoundWords helper so the /results and /compare routes go
  through the service rather than reimplementing the per-word scoring at
  the route boundary (per server/CLAUDE.md).
- Decompose GauntletConfirmPage (287→106), GauntletResultsRoute
  (322→123), and GauntletRoundResultsRoute (310→219) into per-card
  subcomponents under dailyGauntlet/components/. Shared BackButton
  replaces the duplicate inside GauntletHubPage.
- Factor aggregateFromRoundRanks out of getGauntletAggregate so the
  rank-sum / tiebreak / dense-rank logic can be unit-tested without a
  database; add 7 cases covering partial-completion exclusion, ordering,
  best-single tiebreak, last-finish tiebreak, dense ties, latest-finish
  selection, and empty input.
- Collapse a dead ternary in GauntletBottomActions.
