+++ engine/gauntletScoring.ts
@@ -1,10 +1,11 @@
-import type { GauntletModifier } from 'models/gauntlet';
+import { HOT_LETTER_MULTIPLIER, type GauntletModifier } from 'models/gauntlet';
 import { scoreWord } from './scoring.js';
 
 // Per-word score under a gauntlet round's modifier.
 //
 //  - regular:      delegates to the standard length-tier scorer
-//  - hotLetter:    base × multiplier when the word contains the hot letter
+//  - hotLetter:    base × HOT_LETTER_MULTIPLIER when the word contains the
+//                  hot letter
 //  - rareLetters:  sum of per-letter values, replacing length-based scoring
 //
 // The word string is whatever the player's path joined produced (e.g. a Qu
@@ -18,7 +19,7 @@ export function scoreGauntletWord(word: string, modifier: GauntletModifier): num
     case 'hotLetter': {
       const base = scoreWord(upper);
       const hot = modifier.letter.toUpperCase();
-      return upper.includes(hot) ? Math.floor(base * modifier.multiplier) : base;
+      return upper.includes(hot) ? base * HOT_LETTER_MULTIPLIER : base;
     }
     case 'rareLetters': {
       let sum = 0;
